created taskbar and academics section

+++ assets/js/tasks.js
@@ -35,22 +35,50 @@ function displayTasks() {
 
     taskList.innerHTML = "";
 
+    let completedCount = 0;
+
     tasks.forEach(function(task, index){
+
+        if(task.completed){
+            completedCount++;
+        }
 
         const li = document.createElement("li");
         li.className = "task";
 
         li.innerHTML = `
-            <span>${task.text}</span>
+            <div class="task-left">
 
-            <button onclick="deleteTask(${index})">
-                Delete
-            </button>
+                <input
+                    type="checkbox"
+                    ${task.completed ? "checked" : ""}
+                    onchange="toggleTask(${index})">
+
+                <span class="${task.completed ? "completed" : ""}">
+                    ${task.text}
+                </span>
+
+            </div>
+
+            <div>
+
+                <button onclick="editTask(${index})">
+                    Edit
+                </button>
+
+                <button onclick="deleteTask(${index})">
+                    Delete
+                </button>
+
+            </div>
         `;
 
         taskList.appendChild(li);
 
     });
+
+    document.getElementById("progressText").textContent =
+        `Completed ${completedCount} of ${tasks.length} Tasks`;
 
 }
 
@@ -58,6 +86,35 @@ function displayTasks() {
 function deleteTask(index){
 
     tasks.splice(index,1);
+
+    localStorage.setItem("tasks", JSON.stringify(tasks));
+
+    displayTasks();
+
+}
+function toggleTask(index){
+
+    tasks[index].completed = !tasks[index].completed;
+
+    localStorage.setItem("tasks", JSON.stringify(tasks));
+
+    displayTasks();
+
+}
+function editTask(index){
+
+    const newTask = prompt("Edit your task:", tasks[index].text);
+
+    if(newTask === null){
+        return;
+    }
+
+    if(newTask.trim() === ""){
+        alert("Task cannot be empty.");
+        return;
+    }
+
+    tasks[index].text = newTask.trim();
 
     localStorage.setItem("tasks", JSON.stringify(tasks));
